user_dashboard: refactor chain store and fix enabled chain sorting
