+++ apps/user_dashboard/src/state/chains.ts
@@ -4,15 +4,11 @@
  * Chain data is managed by TanStack Query in hooks/queries/use_chains.ts
  */
 
-import { ChainIdHelper } from "@keplr-wallet/cosmos";
 import type { AuthType } from "@oko-wallet/oko-types/auth";
 import { create } from "zustand";
 import { createJSONStorage, persist } from "zustand/middleware";
 
-import type {
-  CosmosChainInfo,
-  ModularChainInfo,
-} from "@oko-wallet-user-dashboard/types/chain";
+import { getChainIdentifier } from "@oko-wallet-user-dashboard/utils/chain";
 
 const STORAGE_KEY = "oko:user_dashboard:chains";
 export const DEFAULT_ENABLED_CHAINS = [
@@ -22,52 +18,7 @@ export const DEFAULT_ENABLED_CHAINS = [
   "osmosis",
 ] as const;
 
-// Cache for ChainIdHelper.parse() results
-const chainIdentifierCache = new Map<string, string>();
-
-/**
- * Get chain identifier with caching to avoid repeated parsing
- */
-export function getChainIdentifier(chainId: string): string {
-  let identifier = chainIdentifierCache.get(chainId);
-  if (!identifier) {
-    identifier = ChainIdHelper.parse(chainId).identifier;
-    chainIdentifierCache.set(chainId, identifier);
-  }
-  return identifier;
-}
-
 type UserKey = `${AuthType}/${string}`;
-const ANONYMOUS_USER_KEY = "__anonymous__";
-
-function createUserKey(authType: AuthType, email: string): UserKey {
-  return `${authType}/${email.trim()}`;
-}
-
-function resolveKey(state: ChainPreferencesState): string {
-  return state.activeUserKey ?? ANONYMOUS_USER_KEY;
-}
-
-function getEnabledChains(state: ChainPreferencesState): string[] {
-  return (
-    state.enabledChainsByUser[resolveKey(state)] ?? [...DEFAULT_ENABLED_CHAINS]
-  );
-}
-
-function updateEnabledChains(
-  state: ChainPreferencesState,
-  updater: (chains: Set<string>) => void,
-): Pick<ChainPreferencesState, "enabledChainsByUser"> {
-  const key = resolveKey(state);
-  const chains = new Set(getEnabledChains(state));
-  updater(chains);
-  return {
-    enabledChainsByUser: {
-      ...state.enabledChainsByUser,
-      [key]: Array.from(chains),
-    },
-  };
-}
 
 interface ChainPreferencesState {
   enabledChainsByUser: Record<string, string[]>;
@@ -81,6 +32,38 @@ interface ChainPreferencesActions {
   disableChains: (...chainIds: string[]) => void;
   isChainEnabled: (chainId: string) => boolean;
   getEnabledChainIds: () => string[];
+}
+
+function createUserKey(authType: AuthType, email: string): UserKey {
+  return `${authType}/${email.trim()}`;
+}
+
+function getEnabledChains(state: ChainPreferencesState): string[] {
+  if (!state.activeUserKey) {
+    return [...DEFAULT_ENABLED_CHAINS];
+  }
+  return (
+    state.enabledChainsByUser[state.activeUserKey] ?? [
+      ...DEFAULT_ENABLED_CHAINS,
+    ]
+  );
+}
+
+function updateEnabledChains(
+  state: ChainPreferencesState,
+  updater: (chains: Set<string>) => void,
+): Partial<Pick<ChainPreferencesState, "enabledChainsByUser">> {
+  if (!state.activeUserKey) {
+    return {};
+  }
+  const chains = new Set(getEnabledChains(state));
+  updater(chains);
+  return {
+    enabledChainsByUser: {
+      ...state.enabledChainsByUser,
+      [state.activeUserKey]: Array.from(chains),
+    },
+  };
 }
 
 export const useChainStore = create<
@@ -133,60 +116,3 @@ export const useChainStore = create<
     },
   ),
 );
-
-export function transformKeplrChain(chain: CosmosChainInfo): ModularChainInfo {
-  const isSVM = !!chain.svm;
-  const isCosmos = !!chain.bech32Config;
-  const isEVM = !!chain.evm;
-
-  const base = {
-    chainId: chain.chainId,
-    chainName: chain.chainName,
-    chainSymbolImageUrl: chain.chainSymbolImageUrl,
-    isTestnet: chain.isTestnet,
-  };
-
-  if (isSVM) {
-    return {
-      ...base,
-      svm: {
-        rpc: chain.svm!.rpc,
-        currencies: chain.currencies,
-      },
-    };
-  }
-
-  if (isCosmos) {
-    return {
-      ...base,
-      isNative: true,
-      cosmos: chain,
-      evm: isEVM
-        ? {
-            chainId: chain.evm!.chainId,
-            rpc: chain.evm!.rpc,
-            currencies: chain.currencies,
-            feeCurrencies: chain.feeCurrencies,
-            bip44: chain.bip44,
-            features: chain.features,
-          }
-        : undefined,
-    };
-  }
-
-  if (isEVM) {
-    return {
-      ...base,
-      evm: {
-        chainId: chain.evm!.chainId,
-        rpc: chain.evm!.rpc,
-        currencies: chain.currencies,
-        feeCurrencies: chain.feeCurrencies,
-        bip44: chain.bip44,
-        features: chain.features,
-      },
-    };
-  }
-
-  return base;
-}
